Uniform temporary directories (#73)
Now the principal functions have the parameters tmpdir and rmtmp to set the temporary directoriy and the choice to delete /not to delete temporary files. Only two exceptions: create_indices_db() uses a fixed temporary directory (since it is a function intended not to be run by endusers); sen2cor() launched from sen2r() uses a default temporary directory if tempdir is in a SAMBA mountpoint over Windows.