Filter SAFE by footprint
Footprint is retrieved from SAFE metadata (online mode) or from the existing archive (offline mode). This improvement allow avoiding the production of black images in offline mode, which occurred in case an AOI overlapped a tile covered by two orbits, of which one only partially, and the AOI was located in the non-covered portion of the tile.