Move ignorelist in the output path (#271)
Now *_ignorelist.txt and *_cloudlist.txt (placed in the directory of the parameter file) are replaced with an hidden TOML file ".ignorelist.txt" containing the list of basenames of non-produced files and the dates of cloud covered images. This file is placed in the output folder, so to be linked with the produced archive.